refactor(options): split production stitch shell ownership
