+++ tools/report-production-code-shape.mjs
@@ -2,7 +2,11 @@ import { existsSync, readFileSync } from 'node:fs';
 
 const HOTSPOT_PATH = 'docs/production-code-hotspots.md';
 const LOC_LIMITS = new Map([
-  ['src/options/app/productionStitchShell.ts', 900],
+  ['src/options/app/productionStitchShell.ts', 500],
+  ['src/options/app/productionStitchActions.ts', 300],
+  ['src/options/app/productionStitchPersistence.ts', 300],
+  ['src/options/app/productionStitchRenderLifecycle.ts', 300],
+  ['src/options/app/productionStitchShellState.ts', 300],
   ['src/options/widgets/YamlConfigWidget.ts', 450],
   ['src/content/video/prompt.ts', 300],
   ['src/content/clipper/components/dialog.ts', 450],
